+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.35] - 2026-02-06
+
 ### Added
 
 - `pcb new --board` now generates `README.md` and `CHANGELOG.md` files from templates
@@ -286,7 +288,8 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 - Error on invalid type passed to `io()`
 - Format the auto-generated component .zen files
 
-[Unreleased]: https://github.com/diodeinc/pcb/compare/v0.3.34...HEAD
+[Unreleased]: https://github.com/diodeinc/pcb/compare/v0.3.35...HEAD
+[0.3.35]: https://github.com/diodeinc/pcb/compare/v0.3.34...v0.3.35
 [0.3.34]: https://github.com/diodeinc/pcb/compare/v0.3.33...v0.3.34
 [0.3.33]: https://github.com/diodeinc/pcb/compare/v0.3.32...v0.3.33
 [0.3.32]: https://github.com/diodeinc/pcb/compare/v0.3.31...v0.3.32